ADD: synctest

+++ Userland/shell/tests/test_util.c
@@ -32,7 +32,7 @@ uint8_t memcheck(void *start, uint8_t value, uint32_t size){
 }
 
 //Parameters
-int64_t satoi(char* str){
+int satoi(char* str){
   uint64_t i = 0;
   int64_t res = 0;
   int8_t sign = 1;
